+++ changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- BVM Doctor updated to check for the BoxLang home bin folder in the system PATH and alert the user if it's missing and fix it.
+
 ### Fixed
 
 - Updated unzip commands to use `-q` for quiet mode and `-o` to overwrite existing files without prompting.